Mechanisms

Index

RigidBodyDynamics.Mechanism

RigidBodyDynamics.TreeDataStructure.path

RigidBodyDynamics.attach!

RigidBodyDynamics.attach!

RigidBodyDynamics.bodies

RigidBodyDynamics.body_fixed_frame_definition

RigidBodyDynamics.body_fixed_frame_to_body

RigidBodyDynamics.fixed_transform

RigidBodyDynamics.joints

RigidBodyDynamics.maximal_coordinates

RigidBodyDynamics.num_positions

RigidBodyDynamics.num_velocities

RigidBodyDynamics.parse_urdf

RigidBodyDynamics.rand_chain_mechanism

RigidBodyDynamics.rand_floating_tree_mechanism

RigidBodyDynamics.rand_mechanism

RigidBodyDynamics.rand_tree_mechanism

RigidBodyDynamics.reattach!

RigidBodyDynamics.remove_fixed_joints!

RigidBodyDynamics.root_body

RigidBodyDynamics.root_frame

RigidBodyDynamics.submechanism

The Mechanism type

RigidBodyDynamics.Mechanism — Type.

type Mechanism{T<:Number}

A Mechanism represents an interconnection of rigid bodies and joints. Mechanisms store the joint layout and inertia parameters, but no state-dependent information.

source

Creating and modifying Mechanisms

RigidBodyDynamics.parse_urdf — Function.

parse_urdf(scalartype, filename)

Create a Mechanism by parsing a URDF file.

source

RigidBodyDynamics.attach! — Function.

attach!(mechanism, predecessor, joint, jointToPredecessor, successor, successorToJoint)

Attach successor to predecessor using joint.

See Joint for definitions of the terms successor and predecessor.

The Transform3Ds jointToPredecessor and successorToJoint define where joint is attached to each body. jointToPredecessor should define joint.frameBefore with respect to any frame fixed to predecessor, and likewise successorToJoint should define any frame fixed to successor with respect to joint.frameAfter.

predecessor is required to already be among the bodies of the Mechanism.

If successor is not yet a part of the Mechanism, it will be added to the Mechanism. Otherwise, the joint will be treated as a non-tree edge in the Mechanism, effectively creating a loop constraint that will be enforced using Lagrange multipliers (as opposed to using recursive algorithms).

source

RigidBodyDynamics.attach! — Method.

attach!(mechanism, parentBody, childMechanism)

Attach childMechanism to mechanism.

Essentially replaces the root body of childMechanism with parentBody (which belongs to mechanism).

Currently doesn't support Mechanisms with cycles.

source

RigidBodyDynamics.maximal_coordinates — Method.

maximal_coordinates(mechanism)

Return a dynamically equivalent Mechanism, but with a flat tree structure with all bodies attached to the root body with a quaternion floating joint, and with the 'tree edge' joints of the input Mechanism transformed into non-tree edge joints (a constraint enforced using Lagrange multipliers in dynamics!).

source

RigidBodyDynamics.rand_chain_mechanism — Method.

rand_chain_mechanism(t, jointTypes)

Create a random chain Mechanism with the given joint types.

source

RigidBodyDynamics.rand_floating_tree_mechanism — Method.

rand_floating_tree_mechanism(t, nonFloatingJointTypes)

Create a random tree Mechanism (without loops), with a quaternion floating joint as the first joint (between the root body and the first non-root body).

source

RigidBodyDynamics.rand_mechanism — Method.

rand_mechanism(?, parentSelector, jointTypes)

Create a random Mechanism with the given joint types. Each new body is attached to a parent selected using the parentSelector function.

source

RigidBodyDynamics.rand_tree_mechanism — Method.

rand_tree_mechanism(t, jointTypes)

Create a random tree Mechanism (without loops).

source

RigidBodyDynamics.reattach! — Function.

reattach!(mechanism, oldSubtreeRootBody, parentBody, joint, jointToParent, newSubtreeRootBody)
reattach!(mechanism, oldSubtreeRootBody, parentBody, joint, jointToParent, newSubtreeRootBody, newSubTreeRootBodyToJoint)

Detach the subtree rooted at oldSubtreeRootBody, reroot it so that newSubtreeRootBody is the new root, and then attach newSubtreeRootBody to parentBody using joint.

Currently doesn't support Mechanisms with cycles.

source

RigidBodyDynamics.remove_fixed_joints! — Method.

remove_fixed_joints!(mechanism)

Remove any fixed joints present as tree edges in mechanism by merging the rigid bodies that these fixed joints join together into bodies with equivalent inertial properties.

Currently doesn't support Mechanisms with cycles.

source

RigidBodyDynamics.submechanism — Method.

submechanism(mechanism, submechanismRootBody)

Create a Mechanism from the subtree of mechanism rooted at submechanismRootBody.

Currently doesn't support Mechanisms with cycles.

source

Basic functionality

RigidBodyDynamics.TreeDataStructure.path — Method.

path(mechanism, from, to)

Return the path from rigid body from to to along edges of the Mechanism's kinematic tree.

source

RigidBodyDynamics.bodies — Method.

bodies(mechanism)

Return the RigidBodys that are part of the Mechanism as an iterable collection.

source

RigidBodyDynamics.fixed_transform — Method.

fixed_transform(mechanism, from, to)

Return the transform from CartesianFrame3D from to to, both of which are rigidly attached to the same RigidBody.

source

RigidBodyDynamics.joints — Method.

joints(mechanism)

Return the Joints that are part of the Mechanism as an iterable collection.

source

RigidBodyDynamics.num_positions — Method.

num_positions(mechanism)

Return the dimension of the joint configuration vector $q$.

source

RigidBodyDynamics.num_velocities — Method.

num_velocities(mechanism)

Return the dimension of the joint velocity vector $v$.

source

RigidBodyDynamics.root_body — Method.

root_body(mechanism)

Return the root (stationary 'world') body of the Mechanism.

source

RigidBodyDynamics.root_frame — Method.

root_frame(mechanism)

Return the default frame of the root body.

source

RigidBodyDynamics.body_fixed_frame_definition — Method.

body_fixed_frame_definition(mechanism, frame)

Return the definition of body-fixed frame frame, i.e., the Transform3D from frame to the default frame of the body to which it is attached.

See also default_frame, frame_definition.

source

RigidBodyDynamics.body_fixed_frame_to_body — Method.

body_fixed_frame_to_body(mechanism, frame)

Return the RigidBody to which frame is attached.

source