Rigid bodies

Rigid bodies

Index

RigidBodyDynamics.RigidBody

RigidBodyDynamics.add_contact_point!

RigidBodyDynamics.add_frame!

RigidBodyDynamics.change_default_frame!

RigidBodyDynamics.contact_points

RigidBodyDynamics.default_frame

RigidBodyDynamics.fixed_transform

RigidBodyDynamics.frame_definition

RigidBodyDynamics.frame_definitions

RigidBodyDynamics.has_defined_inertia

RigidBodyDynamics.is_fixed_to_body

RigidBodyDynamics.spatial_inertia

RigidBodyDynamics.spatial_inertia!

The RigidBody type

RigidBodyDynamics.RigidBody — Type.

type RigidBody{T}

A non-deformable body.

A RigidBody has inertia (represented as a SpatialInertia), unless it represents a root (world) body. A RigidBody additionally stores a list of definitions of coordinate systems that are rigidly attached to it.

source

Functions

RigidBodyDynamics.add_contact_point! — Method.

add_contact_point!(body, point)

Add a new contact point to the rigid body

source

RigidBodyDynamics.add_frame! — Method.

add_frame!(body, transform)

Add a new frame definition to body, represented by a homogeneous transform from the CartesianFrame3D to be added to any other frame that is already attached to body.

source

RigidBodyDynamics.contact_points — Method.

contact_points(body)

Return the contact points attached to the body as an ordered collection.

source

RigidBodyDynamics.default_frame — Method.

default_frame(body)

The CartesianFrame3D with respect to which all other frames attached to body are defined.

See frame_definitions(body), frame_definition(body, frame).

source

RigidBodyDynamics.fixed_transform — Method.

fixed_transform(body, from, to)

Return the transform from CartesianFrame3D from to to, both of which are rigidly attached to body.

source

RigidBodyDynamics.has_defined_inertia — Method.

has_defined_inertia(b)

Whether the body has a defined inertia.

source

RigidBodyDynamics.spatial_inertia! — Method.

spatial_inertia!(body, inertia)

Set the spatial inertia of the body.

source

RigidBodyDynamics.spatial_inertia — Method.

spatial_inertia(b)

Return the spatial inertia of the body. If the inertia is undefined, calling this method will result in an error.

source

RigidBodyDynamics.change_default_frame! — Method.

change_default_frame!(body, new_default_frame)

Change the default frame of body to frame (which should already be among body's frame definitions).

source

RigidBodyDynamics.frame_definition — Method.

frame_definition(body, frame)

Return the Transform3D defining frame (attached to body) with respect to default_frame(body).

Throws an error if frame is not attached to body.

source

RigidBodyDynamics.frame_definitions — Method.

frame_definitions(body)

Return the list of homogeneous transforms (Transform3Ds) that define the coordinate systems attached to body with respect to a single common frame (default_frame(body)).

source

RigidBodyDynamics.is_fixed_to_body — Method.

is_fixed_to_body(body, frame)

Whether frame is attached to body (i.e. whether it is among frame_definitions(body)).

source