Simulation

Simulation

Index

RigidBodyDynamics.OdeIntegrators.OdeResultsSink

RigidBodyDynamics.OdeIntegrators.integrate

RigidBodyDynamics.OdeIntegrators.runge_kutta_4

RigidBodyDynamics.OdeIntegrators.step

RigidBodyDynamics.simulate

Basic simulation

RigidBodyDynamics.simulate — Function.

simulate(state0, finalTime; Δt)

Basic Mechanism simulation: integrate the state from time $0$ to finalTime starting from the initial state state0. Return a Vector of times, as well as Vectors of configuration vectors and velocity vectors at these times.

Uses MuntheKaasIntegrator. See MuntheKaasIntegrator for a lower level interface with more options.

source

Lower level ODE integration interface

RigidBodyDynamics.OdeIntegrators.OdeResultsSink — Type.

abstract OdeResultsSink

Does 'something' with the results of an ODE integration (e.g. storing results, visualizing, etc.). Subtypes must implement:

initialize(sink, state): called with the initial state when integration begins.

process(sink, t, state): called at every integration time step with the current state and time.

source

RigidBodyDynamics.OdeIntegrators.integrate — Method.

integrate(integrator, state0, finalTime, Δt; maxRealtimeRate)

Integrate dynamics from the initial state state0 at time $0$ to finalTime using step size Δt.

source

RigidBodyDynamics.OdeIntegrators.runge_kutta_4 — Method.

runge_kutta_4(scalartype)

Return the Butcher tableau for the standard fourth order Runge-Kutta integrator.

source

RigidBodyDynamics.OdeIntegrators.step — Method.

step(integrator, t, state, Δt)

Take a single integration step.

state must be of a type for which the following functions are defined:

configuration(state), returns the configuration vector in global coordinates;

velocity(state), returns the velocity vector;

additional_state(state), returns the vector of additional states;

set_velocity!(state, v), sets velocity vector to v;

set_additional_state!(state, s), sets vector of additional states to s;

global_coordinates!(state, q0, ϕ), sets global coordinates in state based on local coordinates ϕ centered around global coordinates q0;

local_coordinates!(ϕ, ϕd, state, q0), converts state's global configuration q and velocity v to local coordinates centered around global coordinates q0.

source