URDF parsing and writing

URDF parsing and writing

RigidBodyDynamics.URDF.parse_urdf — Function.

parse_urdf(filename; scalar_type, root_joint_type, remove_fixed_tree_joints)

Create a Mechanism by parsing a URDF file.

Keyword arguments:

scalar_type: the scalar type used to store the Mechanism's kinematic and inertial properties. Default: Float64.

root_joint_type: the joint type used to connect the parsed Mechanism to the world. Default: Fixed{scalar_type}().

remove_fixed_tree_joints: whether to remove any fixed joints present in the kinematic tree using remove_fixed_tree_joints!. Default: true.

source

RigidBodyDynamics.URDF.write_urdf — Function.

Serialize a Mechanism to the URDF file format.

Limitations:

for <link> tags, only the <inertial> tag is written; there is no support for <visual> and <collision> tags.

for <joint> tags, only the <origin>, <parent>, <child>, and <limit> tags are written. There is no support for the <calibration> and <safety_controller> tags.

These limitations are simply due to the fact that Mechanisms do not store the required information to write these tags.

Keyword arguments:

robot_name: used to set the name attribute of the root <robot> tag in the URDF. Default: nothing (name attribute will not be set).

include_root: whether to include root_body(mechanism) in the URDF. If false, joints with root_body(mechanism) as their predecessor will also be omitted. Default: true.

source